Preserve subpixel offsets from spot detection (#1330)
1. Remove the casts to int.

2. Interpolate the coordinates if there are more than one pixel location, otherwise use sel.

3. Update tests.

Fixes #1310

+++ starfish/core/spots/_detect_spots/blob.py
@@ -115,7 +115,7 @@ class BlobDetector(DetectSpotsAlgorithmBase):
         fitted_blobs[Features.SPOT_RADIUS] = converted_radius
 
         # convert the array to int so it can be used to index
-        rounded_blobs = SpotAttributes(fitted_blobs.astype(int))
+        rounded_blobs = SpotAttributes(fitted_blobs)
 
         rounded_blobs.data['intensity'] = measure_spot_intensity(
             data_image, rounded_blobs, self.measurement_function)
